requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.9.0
beautifulsoup4==4.13.4
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.9
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fsspec==2025.3.2
ftfy==6.3.1
google-auth==2.39.0
google-genai==1.11.0
greenlet==3.2.0
h11==0.14.0
honcho==2.0.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
identify==2.6.10
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kombu==5.5.3
markdown2==2.5.3
markdownify==0.13.1
marker-pdf==1.6.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.75.0
opencv-python-headless==4.11.0.86
packaging==25.0
pdftext==0.6.2
pillow==10.4.0
platformdirs==4.3.7
pre_commit==4.2.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==5.0.7
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.2
surya-ocr==0.13.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
websockets==15.0.1
